# internal/docs/swagger.yaml
basePath: /
definitions:
...
  apicontracts.OperatorConfig:
    properties:
      apiVersion:
        type: string
      id:
        type: string
      kind:
        type: string
      spec:
        $ref: '#/definitions/apicontracts.OperatorSpec'
    type: object
  apicontracts.OperatorSpec:
    properties:
      imagePostfix:
        type: string
      tasks:
        items:
          $ref: '#/definitions/apicontracts.OperatorTask'
        type: array
    type: object
  apicontracts.OperatorTask:
    properties:
      index:
        type: integer
      name:
        type: string
      runOnce:
        type: boolean
      version:
        type: string
    type: object
...
  rorerror.RorError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    name: ROR
    url: https://github.com/NorskHelsenett/ror
  description: 'ROR-API, need any help? Go to channel #drift-sdi-devops in norskhelsenett.slack.com
    slack workspace'
  title: Swagger ROR-API
  version: "0.1"
paths:
...
  /v1/operatorconfigs:
    get:
      consumes:
      - application/json
      description: Get all operator configs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apicontracts.OperatorConfig'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get all operator configs
      tags:
      - operatorconfigs
    post:
      consumes:
      - application/json
      description: Create a operator config
      parameters:
      - description: Add a operator config
        in: body
        name: operatorconfig
        required: true
        schema:
          $ref: '#/definitions/apicontracts.OperatorConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apicontracts.OperatorConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Create a operator config
      tags:
      - operatorconfigs
  /v1/operatorconfigs/:id:
    delete:
      consumes:
      - application/json
      description: Delete a operator config by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Delete a operator config
      tags:
      - operatorconfigs
    get:
      consumes:
      - application/json
      description: Get a operator config by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Get a operator config
        in: body
        name: operatorconfig
        required: true
        schema:
          $ref: '#/definitions/apicontracts.OperatorConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.OperatorConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get a operator config
      tags:
      - operatorconfigs
    put:
      consumes:
      - application/json
      description: Update a operator config by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Update operator config
        in: body
        name: operatorconfig
        required: true
        schema:
          $ref: '#/definitions/apicontracts.OperatorConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.OperatorConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Update a operator config
      tags:
      - operatorconfigs
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
  ApiKey:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"